move DOFDescriptor to separate file
